Normalize how Schema is added and add constr. Type
The GeoJSON type was being added to the Schema variable obtained from
mongoose.Schema, but other types where added directly to
mongoose.Schema.Types.xxxx. This has been normalized to
Schema.Types.xxxx

The other addition is the inclusion of the ObjectType to the
constructor of the Schema Element. "GeoJSON.prototype.constructor =
GeoJSON" this will solve the second issue related to parsing/creating JsonSchemas