💣 Breaking change: Replace types approach with sub-documents
Sub-documents can also be used as types, but also be extended as objects,
for example if you want to add additional values as required `properties`
on a Feature that you want to validate.

This solves other problems with prior approach, including that the types
were only being modified on the default mongoose instance and that
the `required` constraint was not being enforced for GeoJSON features.

Issue: #63